NEW_PY/devops_api/requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
alibabacloud-credentials==0.2.0
alibabacloud-ecs20140526==2.1.0
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.1.6
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.3.1
alibabacloud-tea-util==0.3.5
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-ecs==4.24.13
ansible==6.1.0
ansible-core==2.13.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==2.0.10
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
daphne==2.5.0
decorator==5.0.7
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
django-simpleui==2021.11.5
django-smart-selects==1.5.9
djangorestframework==3.12.4
DjangoUeditor==1.8.143
et-xmlfile==1.1.0
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.32.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipython==7.22.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
kubernetes==23.6.0
MarkupSafe==2.0.1
msgpack==1.0.2
multi-key-dict==2.0.3
multidict==6.0.2
numpy==1.22.3
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
paramiko==2.9.2
parso==0.8.2
pbr==5.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyHamcrest==2.0.2
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
python-dateutil==2.8.1
python-jenkins==1.7.0
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.0
resolvelib==0.8.1
rsa==4.7.2
service-identity==21.1.0
simplejson==3.17.6
six==1.14.0
smmap==5.0.0
sqlparse==0.3.1
tencentcloud-sdk-python==3.0.575
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==20.4.1
uritemplate==4.1.1
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.1.0
xlrd==1.2.0
xlwt==1.3.0
yarl==1.7.2
zope.interface==5.1.0
